Add howto guide to downscale a cluster that runs on a single host
The idea is to give readers a low entry point to clustering, from which
they can expand their knowledge.

A Vanilla cluster is a three node cluster that runs on a single host.
In this guide we create the cluster, add data to it, and then remove
two nodes. Some scripts are required, which are available under the
'scripts' folder. The idea is that users download a zip containing
them, or checkout the repo to access them.